config/requirements.txt
Django==2.2.6
protobuf==3.10.0
PyJWT==1.7.1
django-cors-headers
psycopg2
pillow
djangorestframework-jwt
gunicorn==19.3.0
django-cors-middleware
djangorestframework
google-cloud
google-cloud-pubsub==1.0.0